% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell1.R
\docType{data}
\name{cell1}
\alias{cell1}
\title{Time-series of chloroplast accumulation responses in cell 1}
\usage{
data("cell1")
}
\description{
A real dataset containing the time-series of distances of 15 chloroplasts
from a blue microbeam.
}
\details{
A data frames of chloroplast movements (cell1) with 120 rows and 17 columns.
}
\section{cell1}{

\itemize{
\item time : time in minutes, setting the start of a blue microbeam irradiation as 1
\item ex : presence of an explanatory variable (a blue microbeam) represented by 0 or 1
\item chloroplast1 -- chloroplast15 : distances of chloroplasts from a blue
microbeam in micrometer
}
}

\keyword{datasets}
