% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2.R
\docType{data}
\name{cell2}
\alias{cell2}
\title{Time-series of chloroplast accumulation responses in cell 2}
\usage{
data("cell2")
}
\description{
A real dataset containing the time-series of distances of 11 chloroplasts
from a blue microbeam.
}
\details{
A data frames of chloroplast movements (cell2) with 120 rows and 13 columns.
}
\section{Columns}{

\itemize{
\item time : time in minutes, setting the start of a blue microbeam irradiation as 1
\item ex : presence of an explanatory variable (a blue microbeam) represented by 0 or 1
\item chloroplast1 -- chloroplast11 : distances of chloroplasts from a blue
microbeam in micrometer
}
}

\keyword{datasets}
