% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell3.R
\docType{data}
\name{cell3}
\alias{cell3}
\title{Time-series of chloroplast accumulation responses in cell 3}
\usage{
data("cell3")
}
\description{
A real dataset containing the time-series of distances of 13 chloroplasts
from a blue microbeam.
}
\details{
A data frames of chloroplast movements (cell3) with 120 rows and 15 columns.
}
\section{cell3}{

\itemize{
\item time : time in minutes, setting the start of a blue microbeam irradiation as 1
\item ex : presence of an explanatory variable (a blue microbeam) represented by 0 or 1
\item chloroplast1 -- chloroplast13 : distances of chloroplasts from a blue
microbeam in micrometer
}
}

\keyword{datasets}
