% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_dist_beta.R
\name{lm_dist_beta}
\alias{lm_dist_beta}
\title{Robust linear regression (x: distance at time 0, y: coefficient of an explanatory variable or velocity)}
\usage{
lm_dist_beta(
  cell_list,
  mvtime,
  ssm_path,
  ex_sign = "negative",
  ssm_method,
  df_name = "cell",
  res_name,
  ex_name,
  unit1,
  unit2,
  ps = 7,
  theme_plot = "bw"
)
}
\arguments{
\item{cell_list}{(list of data frame) The input time-series data. First column, time (column name "time");
second column, an explanatory variable (0 or 1, column name "ex"); third to the last columns,
distances of cells or organelles from the explanatory variable (
any column names are accepted). See examples below for more details.}

\item{mvtime}{(data frame) The movement time estimated by \link{ssm_individual}
("ssm_individual_mvtime.csv") or \link{ssm_KFAS} ("ssm_KFAS_mvtime.csv").}

\item{ssm_path}{(character string) The path of the output directory of
\link{ssm_individual} defined by the \code{out} parameter when \code{ssm_method} is "Bayes".

The path of the output directory of \link{ssm_KFAS} defined by the \code{out} parameter when \code{ssm_method} is "KFAS".}

\item{ex_sign}{(character string) "positive" or "negative". This is used to
estimate the start time of the positive or negative influence of the explanatory
variable on the distances of cells or organelles.
If cells or organelles are moving away from the explanatory variable, \code{ex_sign}
should be "positive". If cells or organelles are approaching to the explanatory
variable, \code{ex_sign} should be "negative".}

\item{ssm_method}{(character string) "Bayes" or "KFAS".}

\item{df_name}{(character string) The name of the data frame. This is used for
file names and graph titles. The default is "cell".}

\item{res_name}{(character string) The name of a response variable.}

\item{ex_name}{(character string) The name of an explanatory variable.}

\item{unit1}{(character string) The unit of a response variable. One of "meter",
"centimeter", "millimeter", "micrometer", "nanometer". If another character
string is given, it is used as it is. This is used for graph labels.}

\item{unit2}{(character string) The unit of time. This is used for graph labels.}

\item{ps}{(positive integer) Font size of graphs specified in pt. The default is 7 pt.
Plot sizes are automatically adjusted according to the font size.}

\item{theme_plot}{(character string) A ggplot theme. One of "bw", "light",
"classic", "gray", "dark", "test", "minimal" and "void". The default is "bw".}
}
\value{
A list of ggplot objects is returned. In each plot, dots, solid lines
and shaded regions are the observed values, regression lines and 95\% confidence
intervals, respectively.
}
\description{
\code{lm_dist_beta} performs the robust linear regression to analyze the
relationship between the distance at time 0 and the influence of an explanatory variable
obtained by the state-space model.
}
\examples{
### A real data example of chloroplast accumulation responses to a blue microbeam ###

# Load packages
library(cellssm)

# Create an output directory
out <- "05_lm_dist_beta"
if(!file.exists(out)){
  dir.create(out, recursive=TRUE)
}

## Bayes
# Load data
data("cell1", "cell2", "cell3", "cell4", "chloroplast_mvtime")
cell_list <- list(cell1, cell2, cell3, cell4)

# Specify the path of the output directory of [ssm_individual] or [ssm_KFAS]
# below, the path of system files is specified to show an example
ssm_file <- stringr::str_split(system.file("extdata", "individual_model.stan",
                                           package = "cellssm"), "/")[[1]]
ssm_path <- paste(ssm_file[-length(ssm_file)], collapse = "/")

# Linear regression
glist <- lm_dist_beta(cell_list = cell_list, mvtime = chloroplast_mvtime,
                      ssm_path = ssm_path,
                      ssm_method = "Bayes", res_name = "chloroplast",
                      ex_name = "microbeam", unit1 = "micrometer", unit2 = "min")

# Save output
g <- glist[[1]] + glist[[2]] + glist[[3]] + glist[[4]] + glist[[5]] +
  patchwork::plot_layout(nrow = 2)
suppressWarnings(ggplot2::ggsave(paste0(out, "/individual_chloroplast_lm_dist_beta.pdf"),
                                 g, height = 104, width = 168, units = "mm"))

}
