% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_signal.R
\name{lm_signal}
\alias{lm_signal}
\title{Estimation of the signal transfer speed by linear regression (x: start time, y: distance at time 0)}
\usage{
lm_signal(
  cell_list,
  mvtime,
  df_name = "cell",
  ex_name,
  unit1,
  unit2,
  ps = 7,
  theme_plot = "bw"
)
}
\arguments{
\item{cell_list}{(list of data frame) The input time-series data. First column, time (column name "time");
second column, an explanatory variable (0 or 1, column name "ex"); third to the last columns,
distances of cells or organelles from the explanatory variable (
any column names are accepted). See examples below for more details.}

\item{mvtime}{(data frame) The movement time estimated by \link{ssm_individual}
("ssm_individual_mvtime.csv") or \link{ssm_KFAS} ("ssm_KFAS_mvtime.csv").}

\item{df_name}{(character string) The name of data frame. This is used for graph labels.
The default is "cell".}

\item{ex_name}{(character string) The name of an explanatory variable. This is used for graph labels.}

\item{unit1}{(character string) The unit of a response variable. One of "meter",
"centimeter", "millimeter", "micrometer", "nanometer". If another character
string is given, it is used as it is. This is used for graph labels.}

\item{unit2}{(character string) The unit of time. This is used for graph labels.}

\item{ps}{(positive integer) Font size of graphs specified in pt. The default is 7 pt.
Plot sizes are automatically adjusted according to the font size.}

\item{theme_plot}{(character string) A ggplot theme. One of "bw", "light",
"classic", "gray", "dark", "test", "minimal" and "void". The default is "bw".}
}
\value{
A list of ggplot objects is returned. In each plot, dots, solid lines
and shaded regions are the observed values, regression lines and 95\% confidence
intervals, respectively.
}
\description{
\code{lm_signal} estimates the speed of signal transfer by linear regression
between the distance at time 0 and the start time of the
influence of an explanatory variable obtained by the state-space model.
}
\examples{
### A real data example of chloroplast accumulation responses to a blue microbeam ###

# Load packages
library(cellssm)

# Create an output directory
out <- "07_lm_signal"
if(file.exists(out)==FALSE){
  dir.create(out, recursive=TRUE)
}

# Load data
data("cell1", "cell2", "cell3", "cell4", "chloroplast_mvtime")
cell_list <- list(cell1, cell2, cell3, cell4)

# Linear regression
glist <- lm_signal(cell_list = cell_list, mvtime = chloroplast_mvtime,
                   ex_name = "microbeam", unit1 = "micrometer", unit2 = "min")

# Save output
g <- glist[[1]] + glist[[2]] + glist[[3]] + glist[[4]] + glist[[5]] + glist[[6]] +
  patchwork::plot_layout(ncol = 3)

suppressWarnings(ggplot2::ggsave(paste0(out, "/individual_chloroplast_lm_signal.pdf"),
                                 g, height = 110, width = 50*3, units = "mm"))



### A simulated data example of Paramecium escape responses from a laser heating ###

# Load packages
library(cellssm)

# Create an output directory
out <- "17_lm_signal"
if(file.exists(out)==FALSE){
  dir.create(out, recursive=TRUE)
}

# Load data
data("Paramecium", "Paramecium_mvtime")
cell_list <- list(Paramecium)

# Linear regression
glist <- lm_signal(cell_list = cell_list, mvtime = Paramecium_mvtime,
                   df_name = "experiment", ex_name = "heat",
                   unit1 = "millimeter", unit2 = "sec")

# Save output
g <- glist[[1]] + glist[[2]] +
  patchwork::plot_layout(nrow = 1)
suppressWarnings(ggplot2::ggsave(paste0(out, "/individual_Paramecium_lm_signal.pdf"),
                                 g, height = 50, width = 50*1.6, units = "mm"))

}
