% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_KFAS.R
\name{ssm_KFAS}
\alias{ssm_KFAS}
\title{Inference of the state-space model by the Kalman filter}
\usage{
ssm_KFAS(
  cell_list,
  visual = NULL,
  out,
  start_sensitivity = 5,
  ex_sign = "negative",
  df_name = "cell",
  res_name,
  ex_name,
  unit1,
  unit2,
  shade = TRUE,
  start_line = TRUE,
  ps = 7,
  theme_plot = "bw"
)
}
\arguments{
\item{cell_list}{(list of data frame) The input time-series data. First column, time (column name "time");
second column, an explanatory variable (0 or 1, column name "ex"); third to the last columns,
distances of cells or organelles from the explanatory variable (
any column names are accepted). The velocities of the third to the last columns are
calculated from the distance and time, and used as response variables in the
modelling. See the following \strong{Examples} for further details.}

\item{visual}{(data frame) The optional data of visual estimation of the start time of
the influence of an explanatory variable. First column, cells (column name "cell");
second column, index (column name "index"); third column, the start time. The default is \code{NULL}.}

\item{out}{(character string) The path of the output directory.}

\item{start_sensitivity}{(positive integer) The sensitivity to detect the start time
of movements. Larger values indicate a higher sensitivity. The default is 5.}

\item{ex_sign}{(character string) "positive" or "negative". This is used to
estimate the start time of the positive or negative influence of the explanatory
variable on the distances of cells or organelles.
If cells or organelles are moving away from the explanatory variable, \code{ex_sign}
should be "positive". If cells or organelles are approaching to the explanatory
variable, \code{ex_sign} should be "negative".}

\item{df_name}{(character string) The name of the data frame. This is used for
file names and graph titles. The default is "cell".}

\item{res_name}{(character string) The name of the response variable. This is used
for file names and graph labels.}

\item{ex_name}{(character string) The name of the explanatory variable. This is
used for graph labels.}

\item{unit1}{(character string) The unit of the response variable. One of "meter",
"centimeter", "millimeter", "micrometer", "nanometer". If another character
string is given, it is used as it is. This is used for graph labels.}

\item{unit2}{(character string) The unit of time. This is used for graph labels.}

\item{shade}{(logical) Whether to draw shade in graphs during the period without
the explanatory variable. The default is \code{TRUE}.}

\item{start_line}{(logical) Whether to draw a line at the start time of the influence of
the explanatory variable in graphs. The default is \code{TRUE}.}

\item{ps}{(positive integer) Font size of graphs specified in pt. The default is 7 pt.
Plot sizes are automatically adjusted according to the font size.}

\item{theme_plot}{(character string) A plot theme of the \link{ggplot2} package. One of "bw", "light",
"classic", "gray", "dark", "test", "minimal" and "void". The default is "bw".}
}
\value{
A directory named after the \code{out} parameter is created, which has two subdirectories.
\itemize{
\item A subdirectory "csv" includes "ssm_KFAS_cell \code{i} _ \code{res_name} \code{j} .csv",
"ssm_KFAS_sd.csv" and "ssm_KFAS_mvtime.csv" with \code{i} the indexes of cells,
\code{j} the indexes of the response variables. "ssm_KFAS_cell \code{i} _ \code{res_name} \code{j} .csv"
contains the credible intervals of time-varying parameters, where "Y" is
the observed velocity, "alpha" is the true state of velocity, and "b_ex" is
the time-varying coefficient of the explanatory variable. In "ssm_KFAS_sd.csv",
"s_b_ex" and "s_Y" are the system noise and observation error as standard deviations,
respectively. "ssm_KFAS_mvtime.csv" contains the estimated start time,
end time, and period of the directional movement.
\item A subdirectory "pdf" includes "ssm_KFAS_cell \code{i} _ \code{res_name} \code{j} .pdf".
This is the visualised results of the model. The figure consists of three panels:
(1) observed distance of \code{res_name} from \code{ex_name}, (2) velocity of \code{res_name}, and
(3) regression coefficient of \code{ex_name}. In these panels,
dots, solid lines and shaded regions are the observed values, medians and 95\%
credible intervals, respectively. When the optional visual estimation of
the start time is given to the \code{visual} parameter, orange solid lines and
green dashed lines represent the start time estimated by the model and
the visual observation, respectively. When the \code{shade} parameter is \code{TRUE},
the shaded and light regions represent the periods without and with the explanatory
variable, respectively.
}
}
\description{
\code{ssm_KFAS} estimates the parameters of the state-space model to
analyse the velocity of individual cells or organelles by the Kalman filter using the KFAS package.
It estimates the credible intervals of the parameters and the start time of
the influence of an explanatory variable in the regression model.
}
\examples{
### Real data example of chloroplast accumulation responses to a blue microbeam ###

# Load package
library(cellssm)

# Load data of chloroplast movements
data("cell1", "cell2", "cell3", "cell4", "visual")
cell_list <- list(cell1, cell2, cell3, cell4)

# Check the format of input data
cell_list
visual

# Execution of state-space modelling

# When you do not want to compare the statistical and visual estimations of the start time
ssm_KFAS(cell_list = cell_list, out = "03_ssm_KFAS",
         res_name = "chloroplast", ex_name = "microbeam",
         unit1 = "micrometer", unit2 = "min")

# When you do want to compare the statistical and visual estimations of the start time
ssm_KFAS(cell_list = cell_list, visual = visual, out = "03_ssm_KFAS",
         res_name = "chloroplast", ex_name = "microbeam",
         unit1 = "micrometer", unit2 = "min")



### Simulated data example of Paramecium escape responses from laser heating ###

# Load package
library(cellssm)

# Load data
data("Paramecium")
cell_list <- list(Paramecium)

# Check the format of input data
cell_list

# Execution of state-space modelling
ssm_KFAS(cell_list = cell_list, out = "13_ssm_KFAS",
         ex_sign = "positive", df_name = "experiment",
         res_name = "Paramecium", ex_name = "heat",
         unit1 = "millimeter", unit2 = "sec")

}
