% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\docType{data}
\name{visual}
\alias{visual}
\title{Visual estimation of the start time of chloroplast accumulation responses in
four plant cells}
\usage{
data("visual")
}
\description{
A real dataset containing the start time of chloroplast movements
in response to a blue microbeam.
}
\details{
A data frame  (visual) with 50 rows and 3 columns.
}
\section{visual}{

\itemize{
\item cell : index of cells (1--4)
\item index : index of chloroplasts in each cell
\item time : start time of chloroplast movements in minutes
}
}

\keyword{datasets}
