% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellssm-package.R
\docType{package}
\name{cellssm-package}
\alias{cellssm-package}
\alias{cellssm}
\title{cellssm: State-space Modeling for the Directional Movement of Cells}
\description{
This package provides an easy way to analyze the time-series of distances of
cells or organelles from an external stimulus. Main functions are based on the
Bayesian inference of parameters in the state-space model.
}
\details{
This package is composed of main functions to implement state-space modeling
and sub-functions to perform minor tasks.
}
\section{Main functions}{

\itemize{
\item \link{ssm_individual} : Bayesian inference of the state-space model for the individual dynamics
\item \link{ssm_common} : Bayesian inference of the state-space model to extract the common dynamics
\item \link{ssm_KFAS} : Inference of the state-space model for the individual dynamics by the Kalman filter
\item \link{nomodel} : Estimation of the movement without the state-space model
}
}

\section{Sub-functions}{

\itemize{
\item \link{dist_vis} : Visualization of the distance from an explanatory variable
\item \link{lm_dist_beta} : Robust linear regression (x: distance at time 0, y: coefficient of an explanatory variable)
\item \link{lm_dist_start} : Linear regression (x: distance at time 0, y: start time)
\item \link{lm_signal} : Estimation of the signal transfer speed by linear regression (x: start time, y: distance at time 0)
}
}

\section{Datasets}{

\itemize{
\item \link{cell1} : Time-series of chloroplast accumulation responses in cell 1
\item \link{cell2} : Time-series of chloroplast accumulation responses in cell 2
\item \link{cell3} : Time-series of chloroplast accumulation responses in cell 3
\item \link{cell4} : Time-series of chloroplast accumulation responses in cell 4
\item \link{visual} : Visual estimation of the start time of chloroplast accumulation responses in
four plant cells
\item \link{chloroplast_mvtime} : Model estimation of the start and end time of chloroplast accumulation responses in
four plant cells
\item \link{Paramecium} : Time-series of Paramecium escape responses
\item \link{Paramecium_mvtime} : Model estimation of the start and end time of Paramecium escape responses
}
}

\author{
Haruki Nishio, \email{harukin218@gmail.com}, \url{https://orcid.org/0000-0002-6124-782X}

Satoyuki Hirano, \email{mc226592@s.utsunomiya-u.ac.jp}

Yutaka Kodama, \email{kodama@cc.utsunomiya-u.ac.jp}
}
