% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_dist_beta.R
\name{lm_dist_beta}
\alias{lm_dist_beta}
\title{Linear regression (x: distance at time 0, y: coefficient of explanatory variable or velocity)}
\usage{
lm_dist_beta(
  cell_list,
  mvtime,
  ssm_path,
  ex_sign = "negative",
  ssm_method,
  robust = FALSE,
  df_name = "cell",
  res_name,
  ex_name,
  unit1,
  unit2,
  eq_pos = "topleft",
  ps = 7,
  theme_plot = "bw"
)
}
\arguments{
\item{cell_list}{(list of data frame) The input time-series data. First column, time (column name "time");
second column, an explanatory variable (0 or 1, column name "ex"); third to the last columns,
distances of cells or organelles from the explanatory variable (
any column names are accepted). See the following \strong{Examples} for further details.}

\item{mvtime}{(data frame) The movement time estimated by \link{ssm_individual}
("ssm_individual_mvtime.csv") or \link{ssm_KFAS} ("ssm_KFAS_mvtime.csv").}

\item{ssm_path}{(character string) The path of the output directory of
\link{ssm_individual} specified by the \code{out} parameter when \code{ssm_method} is "Bayes".

The path of the output directory of \link{ssm_KFAS} specified by the \code{out} parameter
when \code{ssm_method} is "KFAS".}

\item{ex_sign}{(character string) "positive" or "negative". This is used to
estimate the start time of the positive or negative influence of the explanatory
variable on the distances of cells or organelles.
If cells or organelles are moving away from the explanatory variable, \code{ex_sign}
should be "positive". If cells or organelles are approaching to the explanatory
variable, \code{ex_sign} should be "negative".}

\item{ssm_method}{(character string) "Bayes" or "KFAS".}

\item{robust}{(logical) When \code{TRUE}, performs robust regression using repeated medians
(Siegel regression). When \code{FALSE}, performs linear regression. The default is \code{FALSE}.}

\item{df_name}{(character string) The name of the data frame. This is used for
file names and graph titles. The default is "cell".}

\item{res_name}{(character string) The name of the response variable.}

\item{ex_name}{(character string) The name of the explanatory variable.}

\item{unit1}{(character string) The unit of the response variable. One of "meter",
"centimeter", "millimeter", "micrometer", "nanometer". If another character
string is given, it is used as it is. This is used for graph labels.}

\item{unit2}{(character string) The unit of time. This is used for graph labels.}

\item{eq_pos}{(character string) The position of equations. One of "topleft", "topright",
"bottomleft", "bottomright". The default is "topleft".}

\item{ps}{(positive integer) Font size of graphs specified in pt. The default is 7 pt.
Plot sizes are automatically adjusted according to the font size.}

\item{theme_plot}{(character string) A plot theme of the \link{ggplot2} package. One of "bw", "light",
"classic", "gray", "dark", "test", "minimal" and "void". The default is "bw".}
}
\value{
A list of ggplot objects is returned. In each plot, dots, solid lines
and shaded regions are the observed values, regression lines, and 95\% confidence
intervals, respectively.
}
\description{
\code{lm_dist_beta} performs linear regression to analyse the
relationship between the distance at time 0 and the influence of an explanatory variable
obtained by the state-space model.
}
\examples{
### Real data example of chloroplast accumulation responses to a blue microbeam ###

# Load package
library(cellssm)

## Bayes
# Load data
data("cell1", "cell2", "cell3", "cell4", "chloroplast_mvtime")
cell_list <- list(cell1, cell2, cell3, cell4)

# Specify the path of the output directory of [ssm_individual] or [ssm_KFAS]
# Below, the path of the system files is specified to provide an example
ssm_file <- stringr::str_split(system.file("extdata", "individual_model.stan",
                                           package = "cellssm"), "/")[[1]]
ssm_path <- paste(ssm_file[-length(ssm_file)], collapse = "/")

# Linear regression
glist <- lm_dist_beta(cell_list = cell_list, mvtime = chloroplast_mvtime,
                      ssm_path = ssm_path, ssm_method = "Bayes", robust = TRUE,
                      res_name = "chloroplast", ex_name = "microbeam",
                      unit1 = "micrometer", unit2 = "min")

# Save output
g <- glist[[1]] + glist[[2]] + glist[[3]] + glist[[4]] + glist[[5]] +
  patchwork::plot_layout(nrow = 2)

\dontrun{
# Create an output directory
out <- "05_lm_dist_beta"
if(file.exists(out)==FALSE){
  dir.create(out, recursive=TRUE)
}

# Save output
suppressWarnings(ggplot2::ggsave(paste0(out, "/individual_chloroplast_lm_dist_beta.pdf"),
                                 g, height = 104, width = 168, units = "mm"))
}

}
