% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paramecium.R
\docType{data}
\name{Paramecium}
\alias{Paramecium}
\title{Time-series of \emph{Paramecium} escape responses}
\usage{
data("Paramecium")
}
\description{
A simulated dataset containing the time series of distances of 10 \emph{Paramecium}
from laser heating.
}
\details{
A data frame of \emph{Paramecium} movements with 200 rows and 12 columns.
}
\section{Columns}{

\itemize{
\item time : time in seconds, setting the start of a laser heating as 1
\item ex : presence of an explanatory variable (laser heating) represented by 0 or 1
\item Paramecium1 -- Paramecium10 : distances of \emph{Paramecium} from laser heating
in millimeters
}
}

\keyword{datasets}
